message payload data type changes

+++ src/wrench/simgrid_S4U_util/S4U_Mailbox.cpp
@@ -119,7 +119,7 @@ namespace wrench {
                    mailbox_name.c_str());
       simgrid::s4u::MailboxPtr mailbox = simgrid::s4u::Mailbox::byName(mailbox_name);
       try {
-        mailbox->put(msg, (size_t) msg->payload);
+        mailbox->put(msg, (uint64_t) msg->payload);
       } catch (xbt_ex &e) {
         if ((e.category == network_error) || (e.category == timeout_error)) {
           throw std::shared_ptr<NetworkError>(new NetworkError(NetworkError::SENDING, mailbox_name));
@@ -190,7 +190,7 @@ namespace wrench {
 
       simgrid::s4u::MailboxPtr mailbox = simgrid::s4u::Mailbox::byName(mailbox_name);
       try {
-        comm_ptr = mailbox->put_async(msg, (size_t) msg->payload);
+        comm_ptr = mailbox->put_async(msg, (uint64_t) msg->payload);
       } catch (xbt_ex &e) {
         if (e.category == network_error) {
           throw std::shared_ptr<NetworkError>(new NetworkError(NetworkError::SENDING, mailbox_name));
